requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
asyncpg==0.30.0
python-dotenv==1.0.1
httpx==0.28.1
bcrypt==4.3.0
PyJWT==2.10.1
pydantic[email]==2.10.6
docker==7.1.0
python-multipart==0.0.20
